osd: Initialization of data members
Fixes the coverity issues:

** 717336 Uninitialized scalar field
CID 717336 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member whitelisted is not initialized
in this constructor nor in any functions that it calls.

** 728001 Uninitialized scalar field
CID 728001 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member priority is not initialized
in this constructor nor in any functions that it calls.

** 1405358 Uninitialized scalar field
CID 1405358 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member field read_length.v is
not initialized in this constructor nor in any functions that it calls.

Signed-off-by: Amit Kumar <amitkuma@redhat.com>